ENH: Disable check for framework Python in OSX backend.
Without this check, the osx backend can be used with just `python`
rather than requiring `pythonw`. `pythonw` still works better (window
comes to the front and gives a taskbar icon), but not requiring
`pythonw` is much less surprising to users.

Since we now work with just python, simplify the tests.